pyproject.toml
[project]
name = "smartcs"
version = "0.1.0"
description = "AI 内部微服务引擎 - RAG 问答与会话管理"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    # Database
    "sqlalchemy>=2.0.28",
    "psycopg[binary]>=3.1.18",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    # AI & LLM
    "openai>=1.14.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langsmith>=0.2.0",
    # Vector Database
    "qdrant-client>=1.17.0",
    # Configuration
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    # HTTP Client
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.0",
]
